mbed-cloud-client: bring in upstream cmake class changes
Upstream commit
https://github.com/openembedded/openembedded-core/commit/01245db2893e39ffb5d4a00e4689f048d0698974
changed part of cmake_do_generate_toolchain_file() to fix HOST detection.

We need to mirror that change in the mbed-cloud-client recipe to fix the
build.

NOTE: The idea of overriding these cmake class functions / settings is a
really ugly hack and needs to be fixed in a future commit.

Signed-off-by: Michael Scott <mike@foundries.io>